# swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Api de pokemon",
        "description": "N/A",
        "termsOfService": "N/A",
        "contact": {
            "email": "lcsilva0702@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000/",
            "description": "Server para teste da api"
        }
    ],
    "paths": {
...
        "/pokemon/{name}": {
            "get": {
                "summary": "Retorna pokemons",
                "description": "Esse endpoint retorna o pokemon desejado",
                "Tags": ["Pokemon"],
                "parameters": [{
                    "in": "path",
                    "name": "name",
                    "type": "string",
                    "required": "true",
                    "description": "description of parameter",
                    "schema": {
                        "$ref": "#/components/schemas/searchPokemon"
                    }
                }],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "401": {
                        "description": "Usuario não logado"
                    },
                    "200": {
                        "description": "Pokemon encontrado"
                    },
                    "500": {
                        "description": "Registration Failed"
                    },
                    "400": {
                        "description": "Enter the name of the pokemon you want"
                    }
                }
            }
        }
    }, 
    "components": {
        "schemas": {
...
            "searchPokemon": {
                "type": "string",
                "properties": {
                    "name": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}